remove unnecessary typecasts before calling e8m0_to_f32()

+++ kernels/gemm/a4w4/bench.py
@@ -111,10 +111,10 @@ def torch_reference(a_fp4, b_fp4, a_scales, b_scales, dtype=torch.bfloat16):
 
     # Expand scales: (M, K//32) -> (M, K) via repeat_interleave
     a_scales_f32 = e8m0_to_f32(
-        a_scales.repeat_interleave(SCALE_GROUP_SIZE, dim=1).to(torch.float32)
+        a_scales.repeat_interleave(SCALE_GROUP_SIZE, dim=1)
     )
     b_scales_f32 = e8m0_to_f32(
-        b_scales.repeat_interleave(SCALE_GROUP_SIZE, dim=1).to(torch.float32)
+        b_scales.repeat_interleave(SCALE_GROUP_SIZE, dim=1)
     )
 
     # Scale the dequantized values
